+++ lab1/Lab1Bai1.java
@@ -12,10 +12,9 @@ public class Lab1Bai1 {
         System.out.print("Nhập họ và tên: ");
         String fullName = sc.nextLine();
         System.out.print("Nhập điểm trung bình: ");
-        double Avg = sc.nextDouble();
+        float Avg = sc.nextFloat();
         sc.nextLine();
         sc.close();
         System.out.printf("Sinh viên: %s%nĐiểm trung bình: %.1f điểm", fullName, Avg);
-
     }
 }
